Finally, fixed the bug.
The bug is not actually in any of our jsx file, but it is in index.js file. The way that it gets the root id in newer version is somehow restrict the API's json response, so that is why the Map is not showing the places detail on the card. In order to fix it, we need to use the ReactDOM.render() to get the id of the root. This render() method is deprecated and will be removed in the future. But for now, i will just keep this method. 5 hours of solving, GG.